Simplifications to `think_type` data type conversion

+++ src/objects.rs
@@ -55,7 +55,7 @@ pub struct objdeftype {
 }
 #[no_mangle]
 pub unsafe extern "C" fn initobjects() {
-    objdef[player as libc::c_int as usize].think = playercmd as libc::c_int as u8;
+    objdef[player as libc::c_int as usize].think = playercmd as u8;
     objdef[player as libc::c_int as usize].contact = benign as libc::c_int as u8;
     objdef[player as libc::c_int as usize].solid = true as u8;
     objdef[player as libc::c_int as usize].firstchar = 256;
@@ -66,7 +66,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[player as libc::c_int as usize].hitpoints = 12;
     objdef[player as libc::c_int as usize].damage = 0;
     objdef[player as libc::c_int as usize].points = 0;
-    objdef[goblin as libc::c_int as usize].think = ramstraight as libc::c_int as u8;
+    objdef[goblin as libc::c_int as usize].think = ramstraight as u8;
     objdef[goblin as libc::c_int as usize].contact = monster as libc::c_int as u8;
     objdef[goblin as libc::c_int as usize].solid = true as u8;
     objdef[goblin as libc::c_int as usize].firstchar = (256 + 64) as u16;
@@ -77,7 +77,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[goblin as libc::c_int as usize].hitpoints = 1;
     objdef[goblin as libc::c_int as usize].damage = 1;
     objdef[goblin as libc::c_int as usize].points = 50;
-    objdef[skeleton as libc::c_int as usize].think = ramdiag as libc::c_int as u8;
+    objdef[skeleton as libc::c_int as usize].think = ramdiag as u8;
     objdef[skeleton as libc::c_int as usize].contact = monster as libc::c_int as u8;
     objdef[skeleton as libc::c_int as usize].solid = true as u8;
     objdef[skeleton as libc::c_int as usize].firstchar = (256 + 128) as u16;
@@ -88,7 +88,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[skeleton as libc::c_int as usize].hitpoints = 1;
     objdef[skeleton as libc::c_int as usize].damage = 1;
     objdef[skeleton as libc::c_int as usize].points = 150;
-    objdef[ogre as libc::c_int as usize].think = ramstraight as libc::c_int as u8;
+    objdef[ogre as libc::c_int as usize].think = ramstraight as u8;
     objdef[ogre as libc::c_int as usize].contact = monster as libc::c_int as u8;
     objdef[ogre as libc::c_int as usize].solid = true as u8;
     objdef[ogre as libc::c_int as usize].firstchar = (256 + 67 * 4) as u16;
@@ -99,7 +99,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[ogre as libc::c_int as usize].hitpoints = 5;
     objdef[ogre as libc::c_int as usize].damage = 2;
     objdef[ogre as libc::c_int as usize].points = 250;
-    objdef[gargoyle as libc::c_int as usize].think = gargcmd as libc::c_int as u8;
+    objdef[gargoyle as libc::c_int as usize].think = gargcmd as u8;
     objdef[gargoyle as libc::c_int as usize].contact = monster as libc::c_int as u8;
     objdef[gargoyle as libc::c_int as usize].solid = true as u8;
     objdef[gargoyle as libc::c_int as usize].firstchar = (256 + 67 * 4 + 35 * 9) as u16;
@@ -110,7 +110,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[gargoyle as libc::c_int as usize].hitpoints = 10;
     objdef[gargoyle as libc::c_int as usize].damage = 3;
     objdef[gargoyle as libc::c_int as usize].points = 500;
-    objdef[dragon as libc::c_int as usize].think = dragoncmd as libc::c_int as u8;
+    objdef[dragon as libc::c_int as usize].think = dragoncmd as u8;
     objdef[dragon as libc::c_int as usize].contact = monster as libc::c_int as u8;
     objdef[dragon as libc::c_int as usize].solid = true as u8;
     objdef[dragon as libc::c_int as usize].firstchar = (256 + 67 * 4 + 35 * 9 + 19 * 16) as u16;
@@ -121,7 +121,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[dragon as libc::c_int as usize].hitpoints = 75;
     objdef[dragon as libc::c_int as usize].damage = 5;
     objdef[dragon as libc::c_int as usize].points = 1000;
-    objdef[turbogre as libc::c_int as usize].think = ramstraight as libc::c_int as u8;
+    objdef[turbogre as libc::c_int as usize].think = ramstraight as u8;
     objdef[turbogre as libc::c_int as usize].contact = monster as libc::c_int as u8;
     objdef[turbogre as libc::c_int as usize].solid = true as u8;
     objdef[turbogre as libc::c_int as usize].firstchar = (256 + 67 * 4 + 19 * 9) as u16;
@@ -132,7 +132,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[turbogre as libc::c_int as usize].hitpoints = 5;
     objdef[turbogre as libc::c_int as usize].damage = 2;
     objdef[turbogre as libc::c_int as usize].points = 500;
-    objdef[wallhit as libc::c_int as usize].think = fade as libc::c_int as u8;
+    objdef[wallhit as libc::c_int as usize].think = fade as u8;
     objdef[wallhit as libc::c_int as usize].contact = benign as libc::c_int as u8;
     objdef[wallhit as libc::c_int as usize].solid = true as u8;
     objdef[wallhit as libc::c_int as usize].firstchar = 26;
@@ -143,7 +143,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[wallhit as libc::c_int as usize].hitpoints = 0;
     objdef[wallhit as libc::c_int as usize].damage = 0;
     objdef[wallhit as libc::c_int as usize].points = 0;
-    objdef[dead1 as libc::c_int as usize].think = explode as libc::c_int as u8;
+    objdef[dead1 as libc::c_int as usize].think = explode as u8;
     objdef[dead1 as libc::c_int as usize].contact = benign as libc::c_int as u8;
     objdef[dead1 as libc::c_int as usize].solid = false as u8;
     objdef[dead1 as libc::c_int as usize].firstchar = 29;
@@ -154,7 +154,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[dead1 as libc::c_int as usize].hitpoints = 0;
     objdef[dead1 as libc::c_int as usize].damage = 0;
     objdef[dead1 as libc::c_int as usize].points = 0;
-    objdef[dead2 as libc::c_int as usize].think = fade as libc::c_int as u8;
+    objdef[dead2 as libc::c_int as usize].think = fade as u8;
     objdef[dead2 as libc::c_int as usize].contact = benign as libc::c_int as u8;
     objdef[dead2 as libc::c_int as usize].solid = false as u8;
     objdef[dead2 as libc::c_int as usize].firstchar = (256 + 224) as u16;
@@ -165,7 +165,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[dead2 as libc::c_int as usize].hitpoints = 0;
     objdef[dead2 as libc::c_int as usize].damage = 0;
     objdef[dead2 as libc::c_int as usize].points = 0;
-    objdef[dead3 as libc::c_int as usize].think = fade as libc::c_int as u8;
+    objdef[dead3 as libc::c_int as usize].think = fade as u8;
     objdef[dead3 as libc::c_int as usize].contact = benign as libc::c_int as u8;
     objdef[dead3 as libc::c_int as usize].solid = false as u8;
     objdef[dead3 as libc::c_int as usize].firstchar = (256 + 67 * 4 + 9 * 16) as u16;
@@ -176,7 +176,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[dead3 as libc::c_int as usize].hitpoints = 0;
     objdef[dead3 as libc::c_int as usize].damage = 0;
     objdef[dead3 as libc::c_int as usize].points = 0;
-    objdef[dead4 as libc::c_int as usize].think = fade as libc::c_int as u8;
+    objdef[dead4 as libc::c_int as usize].think = fade as u8;
     objdef[dead4 as libc::c_int as usize].contact = benign as libc::c_int as u8;
     objdef[dead4 as libc::c_int as usize].solid = false as u8;
     objdef[dead4 as libc::c_int as usize].firstchar = (256 + 67 * 4 + 35 * 9 + 16 * 16) as u16;
@@ -187,7 +187,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[dead4 as libc::c_int as usize].hitpoints = 0;
     objdef[dead4 as libc::c_int as usize].damage = 0;
     objdef[dead4 as libc::c_int as usize].points = 0;
-    objdef[dead5 as libc::c_int as usize].think = fade as libc::c_int as u8;
+    objdef[dead5 as libc::c_int as usize].think = fade as u8;
     objdef[dead5 as libc::c_int as usize].contact = benign as libc::c_int as u8;
     objdef[dead5 as libc::c_int as usize].solid = false as u8;
     objdef[dead5 as libc::c_int as usize].firstchar =
@@ -199,7 +199,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[dead5 as libc::c_int as usize].hitpoints = 0;
     objdef[dead5 as libc::c_int as usize].damage = 0;
     objdef[dead5 as libc::c_int as usize].points = 0;
-    objdef[shot as libc::c_int as usize].think = straight as libc::c_int as u8;
+    objdef[shot as libc::c_int as usize].think = straight as u8;
     objdef[shot as libc::c_int as usize].contact = pshot as libc::c_int as u8;
     objdef[shot as libc::c_int as usize].solid = false as u8;
     objdef[shot as libc::c_int as usize].firstchar = 154;
@@ -210,7 +210,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[shot as libc::c_int as usize].hitpoints = 0;
     objdef[shot as libc::c_int as usize].damage = 1;
     objdef[shot as libc::c_int as usize].points = 0;
-    objdef[guns as libc::c_int as usize].think = gunthinks as libc::c_int as u8;
+    objdef[guns as libc::c_int as usize].think = gunthinks as u8;
     objdef[guns as libc::c_int as usize].contact = benign as libc::c_int as u8;
     objdef[guns as libc::c_int as usize].solid = true as u8;
     objdef[guns as libc::c_int as usize].firstchar = (256 + 67 * 4 - 8) as u16;
@@ -221,7 +221,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[guns as libc::c_int as usize].hitpoints = 255;
     objdef[guns as libc::c_int as usize].damage = 0;
     objdef[guns as libc::c_int as usize].points = 0;
-    objdef[gune as libc::c_int as usize].think = gunthinke as libc::c_int as u8;
+    objdef[gune as libc::c_int as usize].think = gunthinke as u8;
     objdef[gune as libc::c_int as usize].contact = benign as libc::c_int as u8;
     objdef[gune as libc::c_int as usize].solid = true as u8;
     objdef[gune as libc::c_int as usize].firstchar = (256 + 67 * 4 - 4) as u16;
@@ -232,7 +232,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[gune as libc::c_int as usize].hitpoints = 255;
     objdef[gune as libc::c_int as usize].damage = 0;
     objdef[gune as libc::c_int as usize].points = 0;
-    objdef[rock as libc::c_int as usize].think = straight as libc::c_int as u8;
+    objdef[rock as libc::c_int as usize].think = straight as u8;
     objdef[rock as libc::c_int as usize].contact = mshot as libc::c_int as u8;
     objdef[rock as libc::c_int as usize].solid = false as u8;
     objdef[rock as libc::c_int as usize].firstchar = 153;
@@ -243,7 +243,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[rock as libc::c_int as usize].hitpoints = 0;
     objdef[rock as libc::c_int as usize].damage = 1;
     objdef[rock as libc::c_int as usize].points = 0;
-    objdef[bigshot as libc::c_int as usize].think = straight as libc::c_int as u8;
+    objdef[bigshot as libc::c_int as usize].think = straight as u8;
     objdef[bigshot as libc::c_int as usize].contact = nukeshot as libc::c_int as u8;
     objdef[bigshot as libc::c_int as usize].solid = false as u8;
     objdef[bigshot as libc::c_int as usize].firstchar = (256 + 192) as u16;
@@ -254,7 +254,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[bigshot as libc::c_int as usize].hitpoints = 0;
     objdef[bigshot as libc::c_int as usize].damage = 1;
     objdef[bigshot as libc::c_int as usize].points = 0;
-    objdef[teleporter as libc::c_int as usize].think = idle as libc::c_int as u8;
+    objdef[teleporter as libc::c_int as usize].think = idle as u8;
     objdef[teleporter as libc::c_int as usize].contact = benign as libc::c_int as u8;
     objdef[teleporter as libc::c_int as usize].solid = false as u8;
     objdef[teleporter as libc::c_int as usize].firstchar = (256 + 236) as u16;
@@ -265,7 +265,7 @@ pub unsafe extern "C" fn initobjects() {
     objdef[teleporter as libc::c_int as usize].hitpoints = 0;
     objdef[teleporter as libc::c_int as usize].damage = 0;
     objdef[teleporter as libc::c_int as usize].points = 0;
-    objdef[secretgate as libc::c_int as usize].think = idle as libc::c_int as u8;
+    objdef[secretgate as libc::c_int as usize].think = idle as u8;
     objdef[secretgate as libc::c_int as usize].contact = benign as libc::c_int as u8;
     objdef[secretgate as libc::c_int as usize].solid = false as u8;
     objdef[secretgate as libc::c_int as usize].firstchar = (256 + 67 * 4 - 12) as u16;
